Propose new sample format.
